layout/base/tests/test_bug1278021.html

<!DOCTYPE HTML>
<html>
<head>
<title>Test for Bug 1278021</title>
<script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
<style type="text/css">
  #parent {
    transform-style: preserve-3d;
  }

  #firstchild {
    width: 100%;
    height: 226px;
    background-color: blue;
  }

  #secondchild {
    display: block;
    position: absolute;
    top: 0px;
    left: 0px;
    width: 100%;
    height: 100%;
    opacity: 0;
  }
</style>
</head>
<body>
<div id="parent">
  <div id="firstchild"></div>
  <div id="secondchild"></div>
</div>

<pre id="test">
<script type="application/javascript">

/** Test for Bug 1278021 **/

is(document.elementFromPoint(50,50), document.getElementById("secondchild"), "Able to hit transformed object");

</script>
</pre>
</body>
</html>
